Step 2
Created Constructor WeatherReport with a one string parameter with the fileName of the file text information.

It reads the relevant city state low and high of the file parts splits them and assigns them to variables then I built a getter to test that the file was reading correctly. No tostring method so only returns address of the Linked lists