Steps 1 and 2
I determined a statistic for my assignment. I decided to track the lowest recorded temperature and count the number of days recorded below freezing for each city.

So I edited my no parameter constructor with the list of objects to exclude the high. I also deleted the references of high from the temperature class for the constructor the class variables and the toString().